requirements.txt
pygame>=2.5.2
customtkinter>=5.2.1
mutagen>=1.47.0
pillow>=10.1.0
mediapipe>=0.10.8
opencv-python>=4.8.1
numpy>=1.24.0
tk>=0.1.0